consider all relevant commands in 'linked_dylibs'
Linked dylibs are not limited to the `LC_LOAD_DYLIB` load command. For
completeness and consistency with `otool -L`, consider all relevant load
commands (all with a `DylibCommand` structure, except `LC_ID_DYLIB`).

Also adjust `change_install_name` to work the same on the modification
side and for consistency with `install_name_tool`.