MachOFile#set_lc_str_in_cmd: Fix Ruby 1.8.7 bug
With Ruby from `homebrew/versions/ruby187` (1.8.7-p374) on OS X 10.11.3
`2**64` unfortunately yields `0` instead of the correct value. Setting
this to the size of the raw data works around that issue and should
equally well, as the size of the entire Mach-O binary should be a sane
upper limit for this.

In practice, the size of a Mach-O binary is already limited to 32 bits
because the elements of `FatArch` and many offsets used in load commands
are only 32 bits wide.