Optimization: Replace control-flow-by-exception with magic testing and incremental reads.
Instead of attempting to open a file with MachOFile.new and catching
a FatBinaryError to attempt FatFile.new, read a few bytes and do some
magic/sanity tests. Once sanity is established, instantiate either a
MachOFile or a FatFile (or error out with a MagicError).

Additionally, stop truncated files (< 4 bytes) ahead of time with
a TruncationError.