Apply an ad-hoc signature after modifying IDs
If we make any changes to a signed binary, we invalidate the signature.
Without resigning, the resulting binary is unusable. To my knowledge,
there's no disadvantage to signing an unsigned binary using an ad-hoc
signature, so we can just do this unconditionally to ensure we don't end
up creating bad binaries. The ad-hoc signature should always be available
and always work, so this should be safe to do in all circumstances.