macho: require `codesign!` call to be done explicitly.
As explained in
https://github.com/Homebrew/brew/pull/9040#issuecomment-723888003
I think it's more desirable to have the `codesign!` calls happen inside
Homebrew/brew instead of inside `macho/tools` methods.

This makes it easier for Homebrew to turn this functionality on/off,
change what platforms are affected, etc. without needing to tag and
vendor a new release each time.