Moved logic out of the constructor
All of the logic in the constructor except for one
runtime check has been moved to the self.field() method.
This in turn calls one of a bunch of def_reader() methods
that act like attr_reader but lazily define variables
internally.

I also made all of the class methods private that are
used by the dsl.